fix: apply PR 403 review follow-ups
